Add mdx steps story
