+++ packages/mini-editor/src/index.tsx
@@ -5,11 +5,21 @@ import {
   MiniEditorHikeProps,
 } from "./mini-editor-hike"
 import { MiniEditor, MiniEditorProps } from "./mini-editor"
-import { EditorTransition } from "./new/editor-transition"
+import { EditorTransition } from "./editor-transition"
+import {
+  FullMiniEditor,
+  FullMiniEditorHike,
+} from "./new-mini-editor"
+import { mdxToStep, mdxToSteps } from "./mdx"
+
 export {
+  mdxToStep,
+  mdxToSteps,
   MiniEditorHike,
   MiniEditor,
   MiniEditorHikeProps,
   MiniEditorProps,
   EditorTransition,
+  FullMiniEditor,
+  FullMiniEditorHike,
 }
