+++ packages/storybook/src/assets/steps.basic.mdx
@@ -1,0 +1,39 @@
+<StepHead>
+
+```js index.js
+console.log(1)
+```
+
+---
+
+```html index.html
+<h1>Hello</h1>
+```
+
+</StepHead>
+
+<StepHead>
+
+```js index.js
+console.log(2)
+```
+
+</StepHead>
+
+<StepHead>
+
+```js bar.js
+console.log(2)
+```
+
+```js foo.js
+console.log(2)
+```
+
+---
+
+```js index.js
+console.log(2)
+```
+
+</StepHead>
