+++ packages/storybook/src/assets/steps.svelte.mdx
@@ -1,0 +1,257 @@
+<Step>
+
+```svelte App.svelte
+<h1>Hello Svelte</h1>
+```
+
+</Step>
+
+<Step>
+
+```svelte App.svelte
+<script>
+  let name = "Svelte"
+</script>
+
+<h1>{name}</h1>
+```
+
+</Step>
+
+<Step>
+
+```svelte App.svelte
+<script>
+  import MyComponent from './MyComponent.svelte';
+</script>
+
+<MyComponent name="Messi" />
+<MyComponent name="Ronaldo" />
+```
+
+</Step>
+
+<Step>
+
+```svelte App.svelte focus=2
+
+```
+
+---
+
+```svelte MyComponent.svelte
+<script>
+  export let name;
+</script>
+
+<div>
+  <button>{name}</button>
+</div>
+```
+
+</Step>
+
+Each svelte file is a different component
+
+<Step>
+
+```svelte MyComponent.svelte focus=2
+<script>
+  export let name;
+</script>
+
+<div>
+  <button>{name}</button>
+</div>
+```
+
+</Step>
+
+You can declare props with the export keyword.
+
+<Step>
+
+```svelte MyComponent.svelte focus=3,8
+<script>
+  export let name;
+  let goalCount = 2;
+</script>
+
+<div>
+  <button>{name}</button>
+  {"⚽".repeat(goalCount)}
+</div>
+```
+
+</Step>
+
+using js inside curly braces
+
+<Step>
+
+```svelte MyComponent.svelte focus=5:7,11[11:32]
+<script>
+  export let name;
+  let goalCount = 2;
+
+  function handleClick() {
+    goalCount = goalCount + 1
+  }
+</script>
+
+<div>
+  <button on:click={handleClick}>{name}</button>
+  {"⚽".repeat(goalCount)}
+</div>
+```
+
+</Step>
+
+adding event handlers
+
+<Step>
+
+```svelte MyComponent.svelte focus=2:3,8,10,11
+<script>
+  import { createEventDispatcher } from 'svelte';
+  const dispatch = createEventDispatcher();
+
+  export let name;
+  let goalCount = 2;
+
+  function handleClick() {
+    goalCount = goalCount + 1
+    dispatch('goal', {player: name})
+  }
+</script>
+
+<div>
+  <button on:click={handleClick}>{name}</button>
+  {"⚽".repeat(goalCount)}
+</div>
+```
+
+</Step>
+
+firing custom events
+
+<Step>
+
+```svelte App.svelte focus=4:6,9[27:42],10[29:44]
+<script>
+  import MyComponent from './MyComponent.svelte';
+
+  function onGoal(event) {
+    console.log(event.detail.player)
+  }
+</script>
+
+<MyComponent name="Messi" on:goal={onGoal}/>
+<MyComponent name="Ronaldo" on:goal={onGoal}/>
+```
+
+</Step>
+
+handling custom events
+
+<Step>
+
+```svelte MyBox.svelte
+<div>
+  something
+</div>
+
+<style>
+  div {
+    border: 8px solid hotpink;
+  }
+</style>
+```
+
+</Step>
+
+adding style
+
+<Step focus="3,9,12">
+
+```svelte App.svelte
+<script>
+  import MyComponent from './MyComponent.svelte'
+  import MyBox from './MyBox.svelte'
+  function onGoal(event) {
+    console.log(event.detail.player)
+  }
+</script>
+
+<MyBox>
+  <MyComponent name="Messi" on:goal={onGoal}/>
+  <MyComponent name="Ronaldo" on:goal={onGoal}/>
+</MyBox>
+```
+
+</Step>
+
+...
+
+<Step focus="2">
+
+```svelte MyBox.svelte
+<div>
+  <slot />
+</div>
+
+<style>
+  div {
+    border: 8px solid hotpink;
+  }
+</style>
+```
+
+</Step>
+
+using slot
+
+<Step focus="4">
+
+```svelte App.svelte
+<script>
+  import MyComponent from './MyComponent.svelte'
+  import MyBox from './MyBox.svelte'
+  const players = ['Messi', 'Ronaldo', 'Laspada']
+  function onGoal(event) {
+    console.log(event.detail.player)
+  }
+</script>
+
+<MyBox>
+  <MyComponent name="Messi" on:goal={onGoal}/>
+  <MyComponent name="Ronaldo" on:goal={onGoal}/>
+</MyBox>
+```
+
+</Step>
+
+rendering lists
+
+<Step focus="4,11,12[18:30],13">
+
+```svelte App.svelte
+<script>
+  import MyComponent from './MyComponent.svelte'
+  import MyBox from './MyBox.svelte'
+  const players = ['Messi', 'Ronaldo', 'Laspada']
+  function onGoal(event) {
+    console.log(event.detail.player)
+  }
+</script>
+
+<MyBox>
+  {#each players as player}
+    <MyComponent name={player} on:goal={onGoal}/>
+  {/each}
+</MyBox>
+```
+
+</Step>
+
+...
